test: Del FE_E2E_TEST_USER_EMAIL env var. Use fixed test email addresses

+++ api/e2etests/e2e_seed_data.py
@@ -1,43 +1,53 @@
-import os
-
+from app.settings.common import (
+    E2E_CHANGE_EMAIL_USER,
+    E2E_SIGNUP_USER,
+    E2E_USER,
+)
 from environments.models import Environment
-from organisations.models import Organisation, OrganisationRole
+from organisations.models import Organisation, OrganisationRole, Subscription
 from projects.models import Project
 from users.models import FFAdminUser
 
 
 class E2ESeedData:
-    fe_e2e_test_user_email = (
-        os.environ["FE_E2E_TEST_USER_EMAIL"]
-        if "FE_E2E_TEST_USER_EMAIL" in os.environ
-        else None
-    )
-    password = "str0ngp4ssw0rd!"
+    def __init__(self) -> None:
+        # User email address used for E2E Signup test
+        self.signup_user: str = E2E_SIGNUP_USER
 
-    def teardown(self):
+        # User email address used for Change email E2E test which is part of invite tests
+        self.change_email_user: str = E2E_CHANGE_EMAIL_USER
+
+        # User email address used for the rest of the E2E tests
+        self.e2e_user: str = E2E_USER
+
+        # Password used by all the test users
+        self.password = "str0ngp4ssw0rd!"
+
+    def teardown(self) -> None:
         # delete users created for e2e test by front end
-        if self.fe_e2e_test_user_email:
-            FFAdminUser.objects.filter(email=self.fe_e2e_test_user_email).delete()
-            FFAdminUser.objects.filter(
-                email=self.fe_e2e_test_user_email + ".io"
-            ).delete()
-            FFAdminUser.objects.filter(email="changeMail@test.com").delete()
+        FFAdminUser.objects.filter(email=self.signup_user).delete()
+        FFAdminUser.objects.filter(email=self.e2e_user).delete()
+        FFAdminUser.objects.filter(email=self.change_email_user).delete()
 
-    def seed_data(self):
+    def seed_data(self) -> None:
         # create user and organisation for e2e test by front end
-        if self.fe_e2e_test_user_email:
-            organisation = Organisation.objects.create(name="Bullet Train Ltd")
-            org_admin = FFAdminUser.objects.create_user(
-                email=self.fe_e2e_test_user_email,
-                password=self.password,
-                username=self.fe_e2e_test_user_email,
-            )
-            org_admin.add_organisation(organisation, OrganisationRole.ADMIN)
+        organisation: Organisation = Organisation.objects.create(
+            name="Bullet Train Ltd"
+        )
+        org_admin: FFAdminUser = FFAdminUser.objects.create_user(
+            email=self.e2e_user,
+            password=self.password,
+            username=self.e2e_user,
+        )
+        org_admin.add_organisation(organisation, OrganisationRole.ADMIN)
 
-            project = Project.objects.create(
-                name="My Test Project", organisation=organisation
-            )
+        project: Project = Project.objects.create(
+            name="My Test Project", organisation=organisation
+        )
+        Environment.objects.create(name="Development", project=project)
+        Environment.objects.create(name="Production", project=project)
 
-            Environment.objects.create(name="Development", project=project)
-
-            Environment.objects.create(name="Production", project=project)
+        # Upgrade organisation seats
+        Subscription.objects.filter(
+            organisation__in=org_admin.organisations.all()
+        ).update(max_seats=2)
